Type-parameterizing the Scala cache. Distinct caches for each geodecoder (Google Maps, Nominating, etc).
The Nominating geodecoder (GeoDecodingProviderNominatim.scala) avoids now
the "match" instruction in Scala, that was receiving warnings from the
compiler that its coverage was not "complete" over all the universe of
possible type-values in the functional-programming sense. It now forces
a "String" value for each of the "country" and of "postal_code", avoiding
the "match" instruction.